game loop

+++ clean.sh
@@ -1,2 +1,13 @@
-rm -rf build
+#! /bin/sh
+
+build_dir="./build"
+bin_dir="./bin"
+
+if [ -d $build_dir ]
+then
+    rm -rf build
+fi
+
 mkdir build
+clear
+echo "Cleaned!"
